// packages/protoc-gen-grpc-js/src/gen/ts/service-all_grpc.ts
export const ServiceAllDefinition: {
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary: MethodDefinition<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.ServerStream
   */
  serverStream: MethodDefinition<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream: MethodDefinition<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.Bidi
   * @deprecated
   */
  bidi: MethodDefinition<StringValue, Int32Value>;
} = {
  unary: {
    path: "/spec.ServiceAll/Unary",
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: StringValue): Buffer => Buffer.from(toBinary(StringValueSchema, value)),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: Int32Value): Buffer => Buffer.from(toBinary(Int32ValueSchema, value)),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  serverStream: {
    path: "/spec.ServiceAll/ServerStream",
    requestStream: false,
    responseStream: true,
    requestSerialize: (value: StringValue): Buffer => Buffer.from(toBinary(StringValueSchema, value)),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: Int32Value): Buffer => Buffer.from(toBinary(Int32ValueSchema, value)),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  clientStream: {
    path: "/spec.ServiceAll/ClientStream",
    requestStream: true,
    responseStream: false,
    requestSerialize: (value: StringValue): Buffer => Buffer.from(toBinary(StringValueSchema, value)),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: Int32Value): Buffer => Buffer.from(toBinary(Int32ValueSchema, value)),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
  bidi: {
    path: "/spec.ServiceAll/Bidi",
    requestStream: true,
    responseStream: true,
    requestSerialize: (value: StringValue): Buffer => Buffer.from(toBinary(StringValueSchema, value)),
    requestDeserialize: (bytes: Buffer): StringValue => fromBinary(StringValueSchema, bytes),
    responseSerialize: (value: Int32Value): Buffer => Buffer.from(toBinary(Int32ValueSchema, value)),
    responseDeserialize: (bytes: Buffer): Int32Value => fromBinary(Int32ValueSchema, bytes),
  },
};

/**
 * @generated from service spec.ServiceAll
 */
export interface ServiceAllServer extends UntypedServiceImplementation {
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary: handleUnaryCall<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.ServerStream
   */
  serverStream: handleServerStreamingCall<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream: handleClientStreamingCall<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.Bidi
   * @deprecated
   */
  bidi: handleBidiStreamingCall<StringValue, Int32Value>;
}

/**
 * @generated from service spec.ServiceAll
 */
export class ServiceAllClient extends Client {
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary(request: StringValue, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary(request: StringValue, metadata: Metadata, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary(request: StringValue, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  /**
   * @generated from rpc spec.ServiceAll.Unary
   */
  unary(request: StringValue, metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientUnaryCall;
  unary(request: StringValue, metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Int32Value>, optionsOrCallback?: CallOptions | requestCallback<Int32Value>, callback?: requestCallback<Int32Value>): ClientUnaryCall {
    return this.makeUnaryRequest(
      ServiceAllDefinition.unary.path,
      ServiceAllDefinition.unary.requestSerialize,
      ServiceAllDefinition.unary.responseDeserialize,
      request,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Int32Value>,
    );
  }
  /**
   * @generated from rpc spec.ServiceAll.ServerStream
   */
  serverStream(request: StringValue, metadata?: Metadata, options?: CallOptions): ClientReadableStream<Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.ServerStream
   */
  serverStream(request: StringValue, options?: CallOptions): ClientReadableStream<Int32Value>;
  serverStream(request: StringValue, metadataOrOptions?: Metadata | CallOptions, options?: CallOptions): ClientReadableStream<Int32Value> {
    return this.makeServerStreamRequest(
      ServiceAllDefinition.serverStream.path,
      ServiceAllDefinition.serverStream.requestSerialize,
      ServiceAllDefinition.serverStream.responseDeserialize,
      request,
      metadataOrOptions as Metadata,
      options,
    );
  }
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream(callback: requestCallback<Int32Value>): ClientWritableStream<StringValue>;
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, callback: requestCallback<Int32Value>): ClientWritableStream<StringValue>;
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream(options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<StringValue>;
  /**
   * @generated from rpc spec.ServiceAll.ClientStream
   */
  clientStream(metadata: Metadata, options: CallOptions, callback: requestCallback<Int32Value>): ClientWritableStream<StringValue>;
  clientStream(metadataOrOptionsOrCallback?: Metadata | CallOptions | requestCallback<Int32Value>, optionsOrCallback?: CallOptions | requestCallback<Int32Value>, callback?: requestCallback<Int32Value>): ClientWritableStream<StringValue> {
    return this.makeClientStreamRequest(
      ServiceAllDefinition.clientStream.path,
      ServiceAllDefinition.clientStream.requestSerialize,
      ServiceAllDefinition.clientStream.responseDeserialize,
      metadataOrOptionsOrCallback as Metadata,
      optionsOrCallback as CallOptions,
      callback as requestCallback<Int32Value>,
    );
  }
  /**
   * @generated from rpc spec.ServiceAll.Bidi
   * @deprecated
   */
  bidi(metadata?: Metadata, options?: CallOptions): ClientDuplexStream<StringValue, Int32Value>;
  /**
   * @generated from rpc spec.ServiceAll.Bidi
   * @deprecated
   */
  bidi(options?: CallOptions): ClientDuplexStream<StringValue, Int32Value>;
  bidi(metadataOrOptions?: Metadata | CallOptions, options?: CallOptions): ClientDuplexStream<StringValue, Int32Value> {
    return this.makeBidiStreamRequest(
      ServiceAllDefinition.bidi.path,
      ServiceAllDefinition.bidi.requestSerialize,
      ServiceAllDefinition.bidi.responseDeserialize,
      metadataOrOptions as Metadata,
      options,
    );
  }
}